Fix -Wfloat-conversion in waLBerla code

+++ src/core/grid_based_algorithms/lbboundaries/LBBoundary.hpp
@@ -76,10 +76,12 @@ public:
 #ifdef EK_BOUNDARIES // TODO: ugly. Better would be a class EKBoundaries,
                      // deriving from LBBoundaries, but that requires completely
                      // different initialization infrastructure.
-  void set_charge_density(float charge_density) {
-    m_charge_density = charge_density;
+  void set_charge_density(double charge_density) {
+    m_charge_density = static_cast<float>(charge_density);
   }
-  void set_net_charge(float net_charge) { m_net_charge = net_charge; }
+  void set_net_charge(double net_charge) {
+    m_net_charge = static_cast<float>(net_charge);
+  }
 
   float &charge_density() { return m_charge_density; }
   float &net_charge() { return m_net_charge; }
